+++ Drivers/AT49BV322DT_Flash/AT49BV322DT_Flash.cpp
@@ -25,8 +25,8 @@ typedef void(*IAP)(uint32_t[], uint32_t[]);
 
 #define AT49BV322DT_FLASH_SECTOR_NUM   (13)
 
-uint32_t g_AT49BV322DT_Flash_SectorAddress[AT49BV322DT_FLASH_SECTOR_NUM];
-uint32_t g_AT49BV322DT_Flash_SectorSize[AT49BV322DT_FLASH_SECTOR_NUM];
+uint64_t g_AT49BV322DT_Flash_SectorAddress[AT49BV322DT_FLASH_SECTOR_NUM];
+size_t g_AT49BV322DT_Flash_SectorSize[AT49BV322DT_FLASH_SECTOR_NUM];
 
 TinyCLR_Result AT49BV322DT_Flash_Read(uint32_t address, size_t length, uint8_t* buffer) {
     DISABLE_INTERRUPTS_SCOPED(irq);
@@ -144,7 +144,7 @@ TinyCLR_Result AT49BV322DT_Flash_GetBytesPerSector(uint32_t address, int32_t& si
     return TinyCLR_Result::Success;
 }
 
-TinyCLR_Result AT49BV322DT_Flash_GetSectorMap(const uint32_t*& addresses, const uint32_t*& sizes, size_t& count) {
+TinyCLR_Result AT49BV322DT_Flash_GetSectorMap(const uint64_t*& addresses, const size_t*& sizes, size_t& count) {
     for (auto i = 0; i < AT49BV322DT_FLASH_SECTOR_NUM; i++) {
         g_AT49BV322DT_Flash_SectorAddress[i] = (i * AT49BV322DT_FLASH_SECTOR_SIZE) | AT49BV322DT_FLASH_BASE_ADDRESS;
         g_AT49BV322DT_Flash_SectorSize[i] = AT49BV322DT_FLASH_SECTOR_SIZE;
